src/bigchaindb_client/requirements.txt
bigchaindb_driver
grpcio
protobuf
